add command for run project
