+++ main_project/modules/AddressBook_m/birthday_m.py
@@ -1,4 +1,4 @@
-from ..Common_m.field import Field
+from main_project.modules.Common_m.field import Field
 from datetime import datetime
 import re
 from rich.text import Text
