+++ main_project/modules/AddressBook_m/name_m.py
@@ -1,4 +1,4 @@
-from ..Common_m.field import Field
+from main_project.modules.Common_m.field import Field
 from rich.text import Text
 
 class Name(Field):
