+++ main_project/modules/AddressBook_m/record_m.py
@@ -1,9 +1,9 @@
-from .name_m import Name
-from .phone_m import Phone
-from .birthday_m import Birthday
-from .email_m import Email
-from .address_m import Address
-from services.errors_wrap import input_error
+from main_project.modules.AddressBook_m.name_m import Name
+from main_project.modules.AddressBook_m.phone_m import Phone
+from main_project.modules.AddressBook_m.birthday_m import Birthday
+from main_project.modules.AddressBook_m.email_m import Email
+from main_project.modules.AddressBook_m.address_m import Address
+from main_project.services.errors_wrap import input_error
 from rich.table import Table
 from rich.console import Console
 from rich.text import Text
