+++ main_project/modules/Common_m/dictionary.py
@@ -1,7 +1,7 @@
-from services.address_book_manager import hello, add_contact, add_phone, add_birthday, show_all_contacts, show_phone, show_birthday, delete_contact, add_email, show_email, add_address, show_address, birthdays_all, birthdays, edit_address, edit_birthday, edit_email, edit_phone, delete_address, delete_birthday, delete_email, delete_phone, show_contact, search_contact, add_name
-from services.note_manager import add_note, show_all_notes, add_tag, delete_tag, show_note, edit_title, delete_note, search_tag, search_message
-from services.commands_logs import all_logs, logs_by_date
-from services.help_manager import help
+from main_project.services.address_book_manager import hello, add_contact, add_phone, add_birthday, show_all_contacts, show_phone, show_birthday, delete_contact, add_email, show_email, add_address, show_address, birthdays_all, birthdays, edit_address, edit_birthday, edit_email, edit_phone, delete_address, delete_birthday, delete_email, delete_phone, show_contact, search_contact, add_name
+from main_project.services.note_manager import add_note, show_all_notes, add_tag, delete_tag, show_note, edit_title, delete_note, search_tag, search_message
+from main_project.services.commands_logs import all_logs, logs_by_date
+from main_project.services.help_manager import help
 
 def command_d(book, notes):
     command_d = {
