+++ main_project/modules/cli.py
@@ -7,7 +7,7 @@ from modules.Common_m.CONSTANT import filename, filenameNotes
 from services.file_manager import load_data, save_data
 
 from services.address_book_manager import hello, add_contact, add_phone, add_birthday, show_all_contacts, show_phone, show_birthday, delete_contact, add_email, show_email, add_address, show_address, birthdays_all, birthdays, edit_address, edit_birthday, edit_email, edit_phone, delete_address, delete_birthday, delete_email, delete_phone, show_contact
-from services.note_manager import add_note, show_all_notes, add_tag, delete_tag, show_note, edit_message, delete_note, search_tag, search_message
+from services.note_manager import add_note, show_all_notes, add_tag, delete_tag, show_note, delete_note, search_tag, search_message
 
 book = AddressBook()
 # notes = Note()
