+++ main_project/services/address_book_manager.py
@@ -1,6 +1,6 @@
-from modules.AddressBook_m.record_m import Record
-from services.file_manager import save_data
-from modules.Common_m.CONSTANT import filename
+from main_project.modules.AddressBook_m.record_m import Record
+from main_project.modules.Common_m.CONSTANT import filename
+from main_project.services.file_manager import save_data
 from rich.table import Table
 from rich.console import Console
 from rich.panel import Panel
