+++ main_project/services/note_manager.py
@@ -1,6 +1,6 @@
-from modules.Notes_m.record_m import Record
-from services.file_manager import save_data
-from modules.Common_m.CONSTANT import filenameNotes
+from main_project.modules.Notes_m.record_m import Record
+from main_project.services.file_manager import save_data
+from main_project.modules.Common_m.CONSTANT import filenameNotes
 from rich.console import Console
 from rich.text import Text
 from rich.table import Table
