+++ main_project/tests/test_address_book_manager_01.py
@@ -5,11 +5,11 @@ project_root = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))
 if project_root not in sys.path:
     sys.path.insert(0, project_root)
 
-from modules.AddressBook_m.addressbook_m import AddressBook
-from modules.Common_m.CONSTANT import filename
-from services.file_manager import load_data, save_data
+from main_project.modules.AddressBook_m.addressbook_m import AddressBook
+from main_project.modules.Common_m.CONSTANT import filename
+from main_project.services.file_manager import load_data, save_data
 
-from services.address_book_manager import hello, add_contact, add_phone, add_birthday, show_all_contacts, show_phone, show_birthday, delete_contact, add_email, show_email, add_address, show_address, birthdays_all, birthdays, edit_address, edit_birthday, edit_email, edit_phone, delete_address, delete_birthday, delete_email, delete_phone
+from main_project.services.address_book_manager import hello, add_contact, add_phone, add_birthday, show_all_contacts, show_phone, show_birthday, delete_contact, add_email, show_email, add_address, show_address, birthdays_all, birthdays, edit_address, edit_birthday, edit_email, edit_phone, delete_address, delete_birthday, delete_email, delete_phone
 
 def main():
     # filename = "testab.pkl"
