+++ main_project/tests/test_note_manager.py
